reorg: delete rcp-kernel, set workspace target to the cpu
Cargo workspace support for tier 3+ targets is lame:
https://honks.io/@seth/111048513387986627

Working around that by setting a target for the entire workspace and by
removing the rcp kernel crate. Also, a single crate for the RCP didn't
make much sense. Being the system's GPU, it will run small shader-esque
applets which will likely by developed in their own crates. I'm planning
a disconnected workspace for RCP applets in the future.

`cargo build` can run (and complete!) in the repo root, but it doesn't build
anything properly because it ignores the per-crate linker scripts. Continue
to use the build shell scripts for ROM assembly.