Reorganize and define hardware interfaces
This is a squash of many commits that worked toward these goals:
 - Abstract hardware using a PAC or similar
 - Rewrite the visual demo to use those abstractions

This commit itself marks a checkpoint; more code infrastructure is
needed before a complete visual demo moves forward. For example, a
heap allocator, macros to work with fixed-point numbers, interrupt
handling, etc.

The project has also been reorganized as follows:

- Renaming of "cpu-kernel" to "game", since its primary responsibility
  is handling interaction with the console and production of visuals.

- Creation of a "kernel" library that will support the game engine.
  It's possible that a third crate will be created in a successor
  commit to house general snippets (that might be useful in shaders,
  themselves housed in separate crates).

In representing the inputs and outputs of registers, a description of
each field is given. Descriptions and names were copied verbatim from
online sources and rewritten/renamed to clarify or cite sources. That
effort has been paused since the scope of this change set was already
quite large. To be continued...